backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepface>=0.0.98",
    "fastapi>=0.128.0",
    "langchain>=1.2.8",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.7",
    "librosa>=0.11.0",
    "numpy>=2.3.5",
    "openai-whisper>=20250625",
    "opencv-python>=4.13.0.92",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyttsx3>=2.99",
    "scikit-learn>=1.8.0",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.5",
    "sqlalchemy>=2.0.46",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "uvicorn>=0.40.0",
]
